Fix all broken unit tests.
We refactored how the modules work, both in how we mount the rendered
result and the sequencing of when render is called (now, on the return
on data).

Notable :
Bringing in reactDom for render tests
Integration 'alternate' in tests don't return the required data
directly, but call onSuccess with that data - as required. This
correctly triggers the render.