Force task refresh when bank-feeds are updated (#2725)
* Force task refresh when bank-feeds are updated

The bank feed tasks are different than most tasks as they rely on current bank-feed status loaded in the BFF, rather than from user interaction. However, user interaction can cause the bank-feed status to change, so we need a way to force app-state to refresh the tasks for new data. Issuing an event will do so, but since no task will be marked as listening to that event, we need a way to bypass the code which filters out useless events, hence the new `force` param to `closeEvent`.

* Update test: result is always boolean