LAIK-33: Use route name for tasks (#3141)
* Use route name for tasks

Rather than hard coding the routes in app-state, we now provide the route name and params and build it on the front end. This provides looser coupling and sidesteps some bugs we were having.

* Expect implicit appcue param for all routes

Appcues can be attached to any page to provide tutorials and walkthroughs, so should universally be allowed.

* Add tests for new constructPath behaviour

`constructPath` now provides region and businessId automatically where it can.