Redux Connect More Components to Read Data from the Store (#3105)
Redux style guide recommendation to prefer having more UI components subscribed to the Redux store and reading data at a more granular level. This typically leads to better UI performance, as fewer components will need to render when a given piece of state changes.

In this small PR, the EmployeeDetailsNzView component is connected to state prop mainTab, and subTab.
When the main or sub tab is selected, the whole view is re-rendered unnecessarily.
The change in this PR is that the mainTab and subTab are not attached to the main view but instead, smaller component are extracted/created to listen to the state that concerns it. This may be insignificant at this point, but in grand scheme of things, state changes and UI re-renders should be considered when building the UI component.